Add armhf build to the GH actrions workflow
When trying to install on older moldels of the Raspberry pi
running 32 bit OS requires a different docker build to produce a
compatible container. Forked and tried to update the workfl;ow so that
it will create an additional Docker image for this and pusgh to my own,
mcdermg, dokcher hub repo.

Also added ignore .swp files to .gitignore as dont want them showing up
while editing.